feat(backend/frontend/shared): implement edit and delete actions to pois ss-495 (#543)
* feat(backend/frontend/shared): implement edit and delete actions to pois ss-434

* refactor(backend): use array destructuring for route name ss-495

* refactor(backend): change code according to the review ss-495

* refactor(backend): change code according to the review ss-495

* refactor(backend): change code according to the review ss-495

+++ apps/backend/src/modules/points-of-interest/points-of-interest.service.ts
@@ -1,4 +1,5 @@
 import { METERS_IN_KM } from "~/libs/constants/constants.js";
+import { configureString } from "~/libs/helpers/helpers.js";
 import { HTTPCode } from "~/libs/modules/http/http.js";
 import {
 	type CollectionResult,
@@ -43,6 +44,32 @@ class PointsOfInterestService implements Service {
 	}
 
 	public async delete(id: number): Promise<boolean> {
+		const poi = await this.pointsOfInterestRepository.findById(id);
+
+		if (!poi) {
+			throw new PointOfInterestError({
+				message: PointOfInterestExceptionMessage.ID_NOT_FOUND,
+				status: HTTPCode.NOT_FOUND,
+			});
+		}
+
+		const details = poi.toDetailsObject();
+
+		if (details.routes.length > 0) {
+			const [route] = details.routes;
+			const { name: routeName } = route as Required<
+				(typeof details)["routes"][number]
+			>;
+
+			throw new PointOfInterestError({
+				message: configureString(
+					PointOfInterestExceptionMessage.CANNOT_DELETE_INCLUDED_IN_ROUTE,
+					{ routeName },
+				),
+				status: HTTPCode.BAD_REQUEST,
+			});
+		}
+
 		const isDeleted = await this.pointsOfInterestRepository.delete(id);
 
 		if (!isDeleted) {
